requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

python-dotenv==1.0.1
requests==2.31.0
selenium==4.20.0
yagmail==0.15.293
